Move Invoices to LegacyInvoices
We need to handle "Legacy" Invoices as well as the new Relationship
Invoices introduced in 2018.
As they have the same name and the same endpoing, the first step is to
rename one of them. I think that the best option is the rename the old
one as this is now a legacy (and will become less and less common).